Fix issue with private key not being submitted
This was related to the misunderstood inclusion of the `DiffSupressionFunc` which
prevented an update to the private key on creation. Add more thorough tests which compare
the return public key. Since the API does not return the private key, we need to compare
new values to the existing resource state.