Handle connector configuration and metadata properly; add refresh token management
Connector configuration is write-only and API will return an empty response when read back.
This is deliberate and will be address in future platform releases.

Refresh tokens can be used to acquire new access tokens which have expired, the parameter
is not required and can be omitted.

Housekeeping, disable automatic changelog generation until fixed